feat: 배치 모듈 분리
