+++ apps/collector-batch/build.gradle.kts
@@ -6,7 +6,6 @@ dependencies {
     // add-ons
     implementation(project(":modules:jpa"))
     implementation(project(":modules:redis"))
-    implementation(project(":modules:kafka"))
     implementation(project(":supports:jackson"))
     implementation(project(":supports:logging"))
     implementation(project(":supports:monitoring"))
@@ -14,15 +13,19 @@ dependencies {
     // web
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-actuator")
-    
+
+    // Spring Batch
+    implementation("org.springframework.boot:spring-boot-starter-batch")
 
     // querydsl
     implementation("com.querydsl:querydsl-apt::jakarta")
 
-    // test-fixtures
+    // test dependencies
+    testImplementation("org.springframework.boot:spring-boot-starter-test")
+    testImplementation("org.springframework.batch:spring-batch-test")
+    testImplementation("com.h2database:h2")
     testImplementation(testFixtures(project(":modules:jpa")))
     testImplementation(testFixtures(project(":modules:redis")))
-    testImplementation(testFixtures(project(":modules:kafka")))
 }
 
 tasks.test {
