+++ apps/collector-batch/src/main/java/com/loopers/domain/entity/ProductMetrics.java
@@ -1,0 +1,41 @@
+package com.loopers.domain.entity;
+
+import com.loopers.domain.BaseEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.Immutable;
+
+import java.time.ZonedDateTime;
+
+@Entity
+@Table(name = "product_metrics")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Immutable
+public class ProductMetrics extends BaseEntity {
+
+    @Column(name = "product_id", nullable = false, unique = true)
+    private Long productId;
+
+    @Column(name = "view_count", nullable = false)
+    private Long viewCount = 0L;
+
+    @Column(name = "like_count", nullable = false)
+    private Long likeCount = 0L;
+
+    @Column(name = "sales_count", nullable = false)
+    private Long salesCount = 0L;
+
+    @Column(name = "total_sales_amount", nullable = false)
+    private Long totalSalesAmount = 0L;
+
+    @Column(name = "last_viewed_at")
+    private ZonedDateTime lastViewedAt;
+
+    @Column(name = "last_liked_at")
+    private ZonedDateTime lastLikedAt;
+}
