+++ apps/collector-batch/src/main/java/com/loopers/domain/ranking/ScoredProductMetrics.java
@@ -1,0 +1,16 @@
+package com.loopers.domain.ranking;
+
+import com.loopers.domain.entity.ProductMetrics;
+
+public record ScoredProductMetrics(ProductMetrics metrics, ProductRankingScore score) {
+
+    public static ScoredProductMetrics of(ProductMetrics metrics) {
+        ProductRankingScore score = ProductRankingScore.calculate(
+                metrics.getViewCount(),
+                metrics.getLikeCount(),
+                metrics.getSalesCount(),
+                metrics.getTotalSalesAmount()
+        );
+        return new ScoredProductMetrics(metrics, score);
+    }
+}
