+++ apps/collector-batch/src/main/java/com/loopers/infrastructure/batch/ProductMetricsJpaRepository.java
@@ -1,0 +1,16 @@
+package com.loopers.infrastructure.batch;
+
+import com.loopers.domain.entity.ProductMetrics;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.time.ZonedDateTime;
+
+@Repository
+public interface ProductMetricsJpaRepository extends JpaRepository<ProductMetrics, Long> {
+
+    Page<ProductMetrics> findByUpdatedAtBetween(ZonedDateTime start, ZonedDateTime end, Pageable pageable);
+
+}
