+++ apps/collector-batch/src/main/resources/application.yml
@@ -1,0 +1,18 @@
+spring:
+  application:
+    name: collector-batch
+  profiles:
+    active: local
+  config:
+    import:
+      - jpa.yml
+      - redis.yml
+      - logging.yml
+      - monitoring.yml
+
+  batch:
+    job:
+      enabled: false
+    jdbc:
+      initialize-schema: never
+
