--- apps/commerce-collector/src/main/java/com/loopers/domain/repository/MonthlyProductRankingRepository.java
@@ -1,10 +1,0 @@
-package com.loopers.domain.repository;
-
-import com.loopers.domain.ranking.MonthlyProductRanking;
-
-import java.util.List;
-
-public interface MonthlyProductRankingRepository {
-    void saveAllRankings(List<MonthlyProductRanking> rankings);
-    void deleteByYearAndMonth(Integer year, Integer month);
-}
