+++ settings.gradle.kts
@@ -4,6 +4,7 @@ include(
     ":apps:commerce-api",
     ":apps:pg-simulator",
     ":apps:commerce-collector",
+    ":apps:collector-batch",
     ":modules:jpa",
     ":modules:redis",
     ":modules:kafka",
